chore: move test utils to __tests__
